Default postCondition Step.completed
Every step now has a condition `completed`, which is by default the post condition. Steps can override this or extend it. But it makes it easier for the majority of steps.

I tried to remove most of these simple postConditions without loosing information. One typical example would be a `ValidatingCondition` or similar.