5.x Get RaftedStatusCheck for all primary databases ()
Aura wants to be able to run the procedure and get the status report for all databases hosted on the server. Currently, they need to run it for all databases. The idea in a nutshell is to change the database input from a string, to a list of strings. We do the check for all databases in that list. If the list is empty, we give the report for all databases locally.

https://github.com/neo-technology/clustering-meta/blob/main/design-docs/crds/0026-rafted-status-check.md