Transaction log initializer use correct transaction id
Before this change, the transaction log initializer would append a transaction matching the last committed one from the metadata store. This is problematic and could lead to transactions with the same id, or the first transaction in the log file to be equal to 1. Now we correctly use the "nextCommittingTransactionId" to fix the above problems.