Add `NoDatabasesReallocated` notification ()
Introduce a notification for when REALLOCATE DATABASES is called, but nothing gets reallocated (as the allocation is already "optimal" with respect to the whatever the allocator deems so, e.g., equal numbers. Now the client will get a nice notification so they'll know nothing was moved and won't get confused when looking at the SHOW DATABASES output afterwards.