Fix ImportCommandTest to be aware of arrays normalization
The ImportCommand enables normalization by default, so the last change
to add support to array normalization broke some assumptions in
ImportCommandTest.

Here we're adding shouldNormalizeArrayTypes() to focus on the properties
with array type. This is very similar to shouldNormalizeTypes(), but we
separate the cases to not make the existing test too hard to read. Also,
we moved to use AssertThat().isSameAs() instead of assertEquals() or
assertSame(), as it handles better the comparisson of array types.

We also fix shouldIgnoreWhitespaceInAndAroundDecimalArrays() and
shouldIgnoreWhitespaceInAndAroundIntegerArrays() to take in
consideration the new normalization behavior. An alternative would be to
disable normalization for these test cases, but since they are not
concerned with normalization - only with arrays - it is better to keep
the default behavior of the command.