Add a custom classloader for procedures ()
This commit is the second take of the procedure classloader that was
orignally introduced in b6a1f86, but
later reverted. The commit addresses three issues that were found in the
original, namely:

 *   components could not be loaded by procedures

 *   the JARs could not be replaced on Windows, and

 *   the classloader would change the behaviour of the DBMS
      even if the feature was disabled.

To make components properly resolvable by the procedure classloader, we
change our approach for resolving classes. We scan all JAR achives to
collect a list of classes that originate from "extensionless" JARs. We
then take steps to preload these classes in our classloader,
circumventing the normal classloader logic. Remaining classes are
classloaded as normal. If the feature is disabled, we retain the original
classloading behaviour.

Second, to address the Windows issue, we now close the classloader on
shutdown. This means that the file handles to the JAR archives are
released, and the files can be replaced. This fact also reveals a
previously undiscovered issue with the feature. Since replacing a JAR is
not possible on Windows, this means that the feature should not be
activated on this platform.

Lastly, the PR brings with it an update of how the test JARs are
created, to ensure that classloading actually occurs from the JAR
archives. Previously, the URLClassLoader in ProcedureJarLoader
would not open the archives, since the classes in the tests were
already on the classpath.