Parametrize main tests in ExtractorsTest
This parametrizes the main value extracting tests in ExtractorsTest,
covering for all existing extractors:
  * Basic value extraction, incl. array version for those that supports
    that.
  * Normalized extractor configuration and value extraction when
    normalize() is called.

This should improve the test coverage for the extractors and help us not
rely on implicit tests of tools that use the extractors under the hood.

Edge cases (e.g. trim) and throw conditions were kept separate, since
they are one-offs, but all of them were converted to assertThat()
syntax.