Refactor ArrayExtractors to remove code duplication
Every implementation of ArrayExtractor contained a duplicate
implementation of the logic for traversing the input string and
extracting each element. This logic is not very trivial nor easy to
follow, so might be very easy to get something wrong or to fix a bug in
one and not in others.

This refactor is based on the observation that all array extractors
actually use an intermediary element type, before converting to the
final output type of the extractor. Based on this, we can use that
intermediary type as a generic argument, provide new abstract methods so
each type implement only its own convertion and extract the duplicated
code and bring it up to the base ArrayExtractor class.