Tidy-up extractors implementation visibility
Since Extractor<T> is the public interface for extractors, we should
close the visibility of all particular implementations so there is no
direct dependency on them.

In order to achieve this we change the methods on Extractors to return
Extractor<Type> instead of the TypeExtractor and provide a new method in
the Extractor interface to allow users of that interface to do dynamic
type checking during runtime, even when we're mostly passing
Extractor<?> around.