Notifications are cached in the AST cache.
We can get deprecation notifications when turning a query string into an AST.
If we don't include the notifications in the cached value in the AST cache,
we will not get those notifications when hitting the AST cache. A query containing
deprecated features should return deprecation notifications always when it is
submitted and not only when it is actually parsed.