Extend CostModel logic for AllNodesScan + Expand
In the CardinalityCostModel we have special logic to make sure that
AllRelationshipScan/RelationshipTypeScan wins over AllNodesScan + Expand.
This logic would previously not kick in, if there are any plans between the
AllNodesScan and the Expand.

However, the logic is also applicable if there are any number of
cardinality preserving unary plans between the AllNodesScan and the Expand.