Teach QuickImport to generate all primitives + arrays
This adds support into DataGeneratorInput to generate all the primitive
types plus all their respective array versions.

The size of the arrays are controlled using two new command line flags,
--min-array-size and --max-array-size, and are piped through the code
via the RandomValues.Configuration to be passed to the RandomValues
generator.

The previous "int" and "long" were changed to produce a int and long
respectively and not be arbitrarily bounded anymore.