Position of USE clause is now checked by semantic analysis.
Historically, awareness of USE clause was quite limited in Cypher stack as USE clause was
stripped away by Fabric layer. Fabric planner also did some semantic-analysis-like checks
like where USE clause can be allowed to be in a query and sub-query.
This change adds those checks to regular semantic analysis. The rules are that USE clause
must always be the first clause in a query or subquery. Importing WITH clause in a sub-query
is the only exception to this rule and is the only clasue that can precede USE.