Refactoring of Query Router transaction
This change splits Query Router transaction from Fabric Transaction.
The main benefit is that Query Router transaction can do with much
simpler concurrency model which greatly simplifies the code and makes
the reasoning about the correctness much simpler.
Another benefit is that Query Router does not throw Fabric Exceptions anymore.

Concurrency model of Query Router transaction:

Transaction termination is the only modification operation that can be invoked
concurrently with other operations. It iterates through all child transactions
and marks them as terminated, which is a thread-safe operation for both kernel
and driver transactions.
Transactions being non-thread-safe apart from termination has been a feature of Kernel
transactions since the dawn of time and all clients of the API are build to respect it.
Therefore since a Query Router transaction is just a thin layer between those clients
and kernel transactions (and driver transactions), it can rely on this concurrency model.