Use LogPosition from channel instead of Files.size()
Files.size() will return the size on disk, which includes the
pre-allocated bytes. When we build a LogPosition manually with it to
pass to .truncate() method, it might deviate from how the code is used
in production.

One of this case is for example the
truncateLogWithCorruptionThatLooksLikePreAllocatedZeros() test, where we
write a random number of zeroes plus a "corrupt byte" to the file, but if
the total number of bytes written is less than what was pre-allocated, than
the test will fail because truncate() will only scan for corruption from
the position we pass to it.

This change should deflake
truncateLogWithCorruptionThatLooksLikePreAllocatedZeros().