Use RandomExtension instead of ThreadLocalRandom
Using RandomExtension for generating random pads of zeroes, allow us to
register the random seed that was used in the test in order to debug
failures in the future.

This also unifies the logic on those two tests, so both are writting a
random quantity of bytes instead one using bytes and other longs (which
produces 4x more bytes written).