Fix FailingDatabasesIT.restartFailedDatabaseWithInstanceRestart for block format
The test is using a custom fs ignoring specific files for some methods but not all. Listing files would find the metadatastore, but checking existence would not. Simplifying the code checking for the marker it instead of adapting the test, as that change makes the production code more readable anyway