db.clearQueryCaches now clears Query Router cache, too
Currently, we just clear everything from the Query Router cache, regardless of the target database. The reason is that this cache contains entries for databases that might not exist locally, so there is no way how to target such entries with the current API.

This should not be a big issue as this cache saves us only query parsing.

We might do something more fine-grained later as a performance optimisation.