Slight rework of PropertySelection
Driver was adding a way to exclude keys from a selection.
The case is an existing entity that has changed properties in tx-state
where the selection given to the store cursor should exclude the overridden
tx-state properties, to not read them unnecessarily from store.

BlockPropertyCursor reads the requested properties eagerly.
This doesn't work well with DefaultPropertyCursor when the entity
exists since before but has changed properties in the tx state,
where the changed properties are read from store and then later ignored
in favor of the tx state. This doesn't happen in record storage since
property values are read more lazily and gets filtered out in
DefaultPropertyCursor#next. Now that filtering moved to PropertySelection
so that both approaches can enjoy it.