Remember all plans that refer to a later deleted variable.
Previously, we only needed to remember the "last"
operator (or operators, in binary plans) that referred to
a variable that would later be deleted. We used those plans
to define the conflicts on.

But, with distinctness, it can happen that the "last"
reference to a variable happens at a point where it is
distinct, making it possible to discard that conflict.
There might be a previous reference, however, where the
column is not distinct. In that case we still need an Eager.

In order to prepare for introducing distinctness into
ConflictFinder, we now remember all plans that refer to
a variable that later gets deleted.