Counts checker reports inconsistency for one more corner case
Found some code that looked to be reporting but actually didn't.
Now we report counts inconsistencies for entities with invalid tokens
that are not present in the counts store at all. This aligns with
the behaviour for when the count store has an entry for the invalid
token but with the wrong count.