Stop passing AccessMode directly to cursors
Currently there is a mix of ways on how different cursors gets the
intended AccessMode to be used for security/access checks: some cursors
requires an AccessMode parameter; some of them accepts for some
initializations but not for others, e.g. requires for scanBatch() but
not for scan() or single(); some read eagerly from Read's
getAccessMode(); other read lazily from Read's getAccessMode().

This also make it more confusing to track what each check is using and
could cause mismatches between what one cursor uses and what one of its
internal cursor might want to use.

This patch removes completely the AccessMode passing for/from cursors
and always relies on the Read's accessMode() method. All the
implementations of this method are thread-safe if correctly usage of
transactions are happening (for example, properly creating independent
ExecutionContext objects for each worker thread).

By cleaning-up all the passing of AccessModes from this part of the API,
it'll also make it more clear that the proper way to make change to
access modes it through the KernelTransaction's overrideWith(context)
call (which is also exposed by the InternalTransaction interface).

Most of the patch is just removing the arguments from the methods and
calls, with some small code reorganization inside cursors to use the
getAccessMode().