More correct max label id
Change max number of label id bits to 31 since that is the maximum the
ints used for the ids can handle. This gives the nicer
TokenCapacityExceededKernelException instead of an ArithmeticException: integer overflow

Technically it should actually be limited to Integer.MAX_VALUE - 8 since that is what
is guaranteed to fit in the TokenRegistry's hashmap on all VMs, but that could be
a problem for anyone that managed to create such a token already.