Add `CordonedServersExistedDuringAllocation` notification ()
This PR introduces the CordonedServersExistedDuringAllocation notification which is returned by any command that triggers an allocation decision and cordoned servers are present in the cluster. Commands that can return this notification are: CREATE DATABASE, RE-SOMETHING DATABASE, ALTER DATABASE, DEALLOCATE DATABASES FROM SERVER and ALTER SERVER. But not REALLOCATE DATABASES as this command has the requirement for no cordoned servers and errors if there are such servers.