Query router inserts processed query into a pre-parser and AST cache of the target database.
Query router needs to parse query strings into ASTs, so it makes sense to populate the caches
of the target database with the result, so it does not have to be done again.

This is not done for System DB queries as it it more trouble than worth. System DB queries might
contain sensitive information and AST cache is not used for them anyway.