Remove fallbackRead from Partitioned*IndexCursor[Scan|Seek]
Now that the old API relying on CursorContext only is gone from
PartitionedScan, those fallback read fields are not used anymore and we
can remove them.

Since the read passed to the constructor is only used at construction
time to check if the current tx has changes, we can also move it outside
and clean-up the constructor too.