Fixes to FailingDatabaseUpgradeTransactionIT
Fixes a problem in FailingDatabaseUpgradeTransactionIT where
the executor service was trying to shut down before the barrier
was released. This will make the test about 10 seconds faster since
all/most of the wait on shutdown of the executor service is avoided.

Also makes the wait on the read tx explicit instead of relying on it
being finished during the 10 seconds wait in the executor service shutdown.