Fixes a multi-vs-single ID allocation issue in IndexedIdGenerator
nextId and nextConsecutiveIdRange doesn't work well if both used for
the same id generator. This is fine right now because that never happens
from the code using IndexedIdGenerator.

There was tho a recent change where nextConsecutiveIdRange would delegate
its call to nextId if numberOfIds==1 and this caused (performance)
problem for multi-id id generators.

It would be great to try and solve the performance issue when both
of these methods are used for the same id generator, but this commit
just disentangles them.