Fixes an issue creating relationship LOOKUP and value index in same tx
A lot of things had to align for this to happen:
 - We're running clustering
 - The db is empty (at least from the storeView's POV)
 - There's currently no relationship type LOOKUP index
 - A single transaction creates both the relationship type LOOK index
   and a relationship value index
 - StorageEngineIndexBehahaviour#useNodeIdsInRelationshipTokenIndex() is true

In a clustered environment the thread applying transactions is a different
one than the one creating it and if storeView says that it's empty
then the index population will be run by the applier thread.
If StorageEngineIndexBehahaviour#useNodeIdsInRelationshipTokenIndex() is true
then relationship type LOOKUP index will be run in a population job
separate from other relationship indexes.

With all that in mind here's the issue: if the LOOKUP index was populated
(run and completed by the applier) before any created relationship value index
in that same transaction, the relationship value index scan would find
that LOOKUP index and try to acquire a shared lock for it, however the thread
that created it is still holding the exclusive lock on it.

The solution of choice in this commit is to ensure that the relationship
LOOKUP index is created after relationship value indexes in the same tx.