Order by expression should not use same variable name on 2 sides of CartesianProduct
One small plan regression because of the changes is that
for queries that have the same expression in ORDER BY
and a later DISTINCT/Aggregation, we don't reuse the column that
Sort introduces for that expression in the DISTINCT/Aggregation
anymore. Cached properties should alleviate any performance
regressions caused by this.