Register created tokens directly when they are applied to store
This is done for two reasons:
1. There was a small window, where the token is committed, before it is added to the TokenRegistry, where we could see a
possible corruption if that thread crashed, while another concurrently accesses that token.
2. In preparation for removing special handling for ReplicatedTokens. There was special application for those in
transaction application already, now all token transactions do the same and there would be double adding for the
replicated token case in the future.

As a side effect any token creation not going through the regular
transaction flow now needs to add these tokens to the registry
themselves.