Fix bug in ConvertToNFA
Only inline a predicate if it actually depends on the entity of the current state.
Otherwise, we would inline predicates with no dependencies into multiple transition predicates.

Also, as a side-effect of the refactoring, disallow inlining IRExpressions in QPPs.
We cannot test this currently, but after ExtractQPPPredicates for SHORTEST
they should work (as nonInlinedPreFilters).