More accurate index population progress
Previously the progress of an index population was based on how many
entities had been fed into the pipeline, where generating index updates
and finally writing them to the index that it's building, is further
down the line in the pipeline.

This would often have the reported progress be ahead of the actual progress,
the amount would depend on number of populating threads and batch size.
For example it would almost always be the case that it would seem that
an index population would start at 10-20% or so (depending on threads
and also dataset size), getting to 100% and then staying there for a
long time before actually completing.