Dynamically adapt commit process to raft group membership size ()
Make single truly great again!

Allows single-member rafted databases to use the single commit process (i.e., non-rafted) until they are, in fact, in a raft group bigger than 1. This increases the performance of single rafted databases because the single commit process can apply changes to the stores in parallel, whereas the replicated can't.

We do this using a special commit process that will be delegated to the appropriate one underneath, respecting the current size of the raft group membership.