Adds GBPTree Writer#yield
For the ability to reduce contention with other parallel writers.
It's particularly beneficial after write calls knowing that the next
write call to it may not be immediate, perhaps a costly or I/O dependent
calculation of the next data to write.

Also adds this call up the stack for IndexUpdater/IndexImporter.