reorder rolling upgrade protections
Slightly reorder the checks for the index prototype to ensure the right
information of issues is presented.

1. is the index provider usable
2. is the schema valid
   a. can the entity type be used
   b. are there repeated tokens
   c. is composite valid
3. has name
4. no schema already exists

Ensuring the name, asserting it's present, then getting it seemed odd.
If for whatever reason we enter this state: asserts can be enabled or
disabled. If enabled, the assert throws. If disabled the Optional::get
throws, but not warned as an Optional::isPresent check has been made (as
they're indended to be used together).

Thus changing this to the more explicit Optional::orElseThrow by itself.

Though this seems like a change in behaviour in the case of enabled
assertions, it is not. Ensuring the name will never set a null name as
SchemaNameUtil::generateName cannot not return null.