Parallel runtime sets ProvidedOrder sometimes now.
Previously, when using parallel runtime, we would never set a ProvidedOrder.
That is not correct though, as some operators (noticeably Sort) do provide order,
even in parallel, and some operators are also guaranteed to propagate order.

To make things worse, leveraged order was set to true anyway, but that machinery was half broken
since it relies on ProvidedOrder being set correctly.

Now, an exception is thrown (not an AssertionError any more), if we encounter that leveraged order
should set to be true but provided order does not agree. This will help identify cases where
parallel runtime currently returns results in the wrong order.