add minimum required kernel version to IndexProviders
The availability of an index provider is determined by the kernel
version. Since all other indexes were part of 5.0, they were assumed to
be supported.

Vector indexes are the only new index since 5.0, and some basic
protections were put in place; however, with the coming addition of
vector-2.0, it would be nice for the index provider itself to be able to
provide whether or not it is supported.

This adds the minimum required kernel verison to support a particular
index provider. This information can be used for rolling upgrade
protections, rather than having many nested conditionals.