Never plan PartialSort/PartialTop/OrderedDistinct/OrderedAggregation in parallel runtime
The planner now supports generating plans that ensure the above operators are only planned
in positions where they are in a serial pipeline that starts with `Sort` or `Top`.
However, the runtime does not have test coverage for these operators and some implementations for ASMs are missing.

We can hopefully enable planning these operators in parallel after the runtime team has made sure it can be done safely.