rename test utility class and api
Years ago I wrote the utility to help generate tokens for the partition
scanned tests, but it was useful to extract and use elsewhere.

I was mistaken at the time thinking that "token" only referred to
"label" and "relationship type", and "property key" was not a token;
thus I called it "Tags".

Property keys are tokens, thus renaming this class.

Also returning a `List<Integer>` is awkward, almost everything using
this would prefer to take an `int[]`, thus changing the API to return an
`int[]`.

Both have been pet peeves since.