use AnyValue in VectorCandidate
Expanding support for more numerical array types.

Be it manually typed vectors, or constructed, cypher can provide a
ListValue (a VirtualValue), which is some wrapper around a container of
AnyValue.

Also manually typed vectors, or ones passed through some
stringifications could be considered a "mixed" list. LIST<INTEGER|FLOAT>
but would coerse to a LIST<FLOAT> on storage.

The usual case is a FloatingPointArray.

The less likely case of IntegralArrays would go via a NumberArray which
will incur a copy via a NumberValue first. This could be slightly
avoided with additions to certain base classes.

The rest would go via a SequenceValue which is an Iterable<AnyValue>.
This will cover all the VirtualValues, and a general catchall for
anything that is vector-like.