Migration across engines shares allocated page cache
neo4j-admin migrate has a `--pagecache` option which controls the
amount of page cache allocated to do the migration.
Previously this page cache was only used for reading the database to
migrate from, letting the writer for the new database to allocate
its own page cache, always with the default (fairly small) page cache.

This had two downsides:
- Migration process allocating more memory than specified
- Writing side potentially getting a smaller than optimal page cache

Now the allocated page cache is shared between reader and writer.