rawTxPull rotates tx log on version change
The append of byte arrays directly to the transaction log wouldn't
do rotation on version change because it didn't know the contents
it was writing. Now each log file sent over knows its kernel version
and we can rotate appropriately when a change is seen.

With this change all cases of tx log getting the version change
will trigger a rotation of the log.
This means that from now on each transaction log file should only
contain transactions of a single kernel version.

This is not true for checkpoint log files yet.