Parallel GBPTreeWriter has even less overhead on un-contended writes
A GBPTree keeps navigated tree path and latches between operations.
Previously a parallel GBPTreeWriter would, with a certain frequency,
force a reset of this state just to be nice and not starve other potential
concurrent writers that need to write-latch the root
(mostly pessimistic operations). This pre-emptive reset of state is unnecessary
if there are no concurrent writers doing these types of operations, or indeed
if there are no concurrent writers at all.

Therefor this commit adds a check to see if there are other concurrent
writers that want to write-latch the root and only then do the
reset of the state between operations. The check is made with a certain
modulo the number of operations, and this modulo is also reduced because it's
generally way cheaper and so that writers can play even nicer with each other.