Follow "virtual" paths in FileUtils.getCanonicalFile ()
In the docker images, the data folder (`/var/lib/neo4j/data`) was
symlinked directly to the root filesystem (`/data`).

When creating the `Neo4jLayout` the following paths (collected from config)

    dataDirectory = /var/lib/neo4j/data
    databaseDirectory = /var/lib/neo4j/data/databases

were in 4.4.27 canonicalized as

    dataDirectory = /data
    databaseDirectory = /var/lib/neo4j/data/databases

and in 4.4.26 canonicalized as

    dataDirectory = /var/lib/neo4j/data
    databaseDirectory = /var/lib/neo4j/data/databases

rather than the expected outcome

    dataDirectory = /data
    databaseDirectory = /data/databases

The observed discrepancy between the two versions can be explained
by the fact that the function checked for existence,
before performing the `toRealPath` call.

Due to an seemingly innocuous change in https://github.com/neo-technology/neo4j/commit/3a7814f412e16abe0c5c4362ee0889d88634c70f
where the following line was added:

    ctx.fs().mkdirs( databaseLayout.databaseDirectory() );

we ended up in a mixed state where `dataDirectory` was created, but not
`databaseDirectory`. This made the `FileUtil.getCanonicalPath` able to
properly resolve one of the two directories, whereas the other simply
got normalized. This resulted in an `IllegalStateException` on
`BatchingNeoStores.cleanup`.

Before said change, none of the paths existed, and thus the resolution
of the paths failed equally for all values.

We also note that 5.15 does not seem to be affected by this bug.
This is due to another subtle interaction of `ParallelBatchImporter`,
which performs a `RecordDatabaseLayout.convert` after the fact that the
files have been created. In 4.4, the passed on `DatabaseLayout` object
is an instance of `RecordDatabaseLayout`, and thus the value is simply
returned. In 5.15, however, the value passed to the method is of the
type `PlainDatabaseLayout`. This prompts the function to create a new
`RecordDatabaseLayout` object, which again performs a
`getCanonicalPath`, but now with files having been created.

This patch, modifies the behaviour of `getCanonicalPath` to resolve
"virtual" paths where only a part of the path exists.

Fixes #13373.