Change event logging in RaftLogShipper
* Reduce the amount of event, revert som of them back to normal logic.
* Use Parameters instead of String.format

I don't think event publishing should replace normal logging. Events should provide the narrative of the statemachine. Therefor events are now published when: stopped, started, mode is changed or when compaction info is sent to the follower.

Most warns and errors are now reverted back to normal log lines. The idea is that these don't actual change the state or aborts the RLS. It just informs that the current operation terminated unexpectedly but the RLS will carry on.