OutOfOrderSequence optimizations for mvcc ()
MVCC database needs to know ids of uncompleted transactions at the start of every transaction. It does it by asking reverse snapshot from OutOfOrderSequence.
This PR adds some improvements to that functionality, without affecting performance of non-mvcc stores (hopefully).

Simple cache for ReverseSnapshot. It cached in the volatile filed and every time new number is offered the field is reset to null. Readers if observe null create and cache snapshot for use by other lucky readers. This reduces number of actual scans for missing items in the sequence.
SequenceArray now keeps track of number of missing numbers, so it can create array of exact size when missing items/or snapshot is requested, thus avoiding extra allocations.
Not entirely related, snapshot of sequence is simplified and doesn't contain meta anymore, because it is never used, also avoiding some extra allocations.