Track max reserved memory in parallel runtime
In the parallel runtime we don't track the total heap usage of the query since
allocation tacking happens in multiple threads and keeping the exact value can be costly.
However, we can fairly cheaply keep track of the maximum number of bytes we have reserved from
the transaction memory pool (we may not use all of it but it is still reserved). This PR adds
a mechanism for keeping track of this and also updates query.log to use this instead of -1 as before.