Allow "asynchronous" CommitEvents to be created without a TransactionEvent ()
With CommitEvent.NULL we lose e.g. metrics that are incremented by the child LogAppendEvent and StoreApplyEvents . On the other hand, we don't always have access to the TransactionEvent for the transaction that was committed. To work around this, this allows creating a CommitEvent without an outer TransactionEvent in such asynchronous ("detached"?) contexts.