Remove replicated tokens ()
Removed the ability to create ReplicatedTokens. Instead use the same token creation strategy as Kernel. This means that tokens are now just replicated transactions.

The state machines are kept but renamed to Legacy. This is because in a roll any instance may still receive ReplicatedTokens during a roll and we must therefor be able to process those as before.