Add START_OFFSET envelope type w/ reader support ()
This adds a new envelope type, START_OFFSET, to represent an offset that should be skipped at the beginning of the first segment of a log file so we can make sure that replicated entries in a clustered environment are correct aligned.

The ReadChannel will enforce that the payload content of the START_OFFSET envelope should be zeros. This is to fail fast on a possible corrupted log file.