change vector field key in vector-2.0
This will allow us to potentially drop the requirement of declaring the
vector dimensions when creating a vector-2.0 index. Done in a way that
could remove the restriction in a future version, if not ready for the
same release as vector-2.0. (additive change)

vector-1.0 will simply keep using:
    "vector"

vector-2.0 will use the dimensionality, thus for a float[1536]:
    "1526d-vector"

We need some form of VectorIndexVersion -> VectorDocumentStructure
mapping. This could be done as an override on VectorIndexVersion, or
some field accessor; however, it is lucene specific. I would like to try
and keep VectorIndexVersion agnostic to the implementation of the index.

Using switch expression as this is effectively the same as implementing
the effective vtable lookup manually thus shouldn't have much additional
cost if any.

Using the exhaustive switch expression will enforce a compile-time
check. It will also generate a TABLESWITCH instruction rather than a
LOOKUPSWITCH instruction; the former being the same or faster than the
latter ...but such micro optimisations don't really matter, just a nice
perk in addition to the compile-time check.