use builder pattern for lucene index writer configs
The IndexConfig only exists in IndexWriterConfigs in order for vector
indexes to handle the correct Codec. It would be nice to lift the Code
out such that we can remove that dependency.

It is easier and nicer to pull IndexConfig things out than to force with
in. A builder pattern helps separate and configure the IndexWriterConfig
at different levels, before building it when needed.

This also allows for more things to be versioned and handled via
VectorIndexVersion; such as the dimensionalilty. As well as allow for
more overloads that the LuceneIndexType enum tried to already reduce.

Initially tried simply using a Mode enum:
        STANDARD, POPULATION, and TRANSACTION_STATE.

Though this worked, it still kept branching on LuceneIndexType enums
within the implementation body, and as it scaled would hurt readability.
TRANSACTION_STATE was also only used for fulltext indexes.

So rather than mode->type, this reverses it to type->mode.
This allows Text, Fulltext, and Vector indexes to have their own modes.
It would be nice if each mode was an enum, explicit singletons were used
as they share settings.

This approach therefore:
* replaces the helper functions and LuceneIndexType
* removes the need for IndexConfig to be plumbed down
* allows for later customisation by defering IndexWriterConfigBuilder::build
* avoids lambda for the often required Suplier<IndexWriterConfig> as
  IndexWriterConfigBuilder::build is already one

From: IndexWriterConfig.population(LuceneIndexType.FULLTEXT, config, analyser, indexConfig);
To:   new IndexWriterConfigBuilder(FulltextModes.POPULATION, config).withAnalyser(analyser).build();