GBPTree gets properly read-only for "permanently" read-only mode
This has already been fixed in 5.x. The problem is that it wasn't possible
to open a GBPTree in read-only mode, i.e. preventing it from doing any type
of change to the file. Now this will happen if passing in a DatabaseReadOnlyChecker
which is "permanently" read-only, like DatabaseReadOnlyChecker#readOnly().