remove explicit secondary sorting on neighbor entity id
Neighbors who's score to a query is the same sometimes were not sorted
as expected. ScoredEntityResultCollector::subordinate only was comparing
scores; however, Neighbor::compareTo was secondarily comparing entity
ids.

It is possible for multiple different vectors to have the same score to
a query, but its easiest to emulate this senario with duplicate vectors.

The comparison for Neighbor is really only used in tests, and does not
reflect what a user would actually see, unless we specifically returned
Stream::sorted.

Initially I tried to make ScoredEntityResultCollector::subordinate care
about entity id; however, initial testing still showed the same problem.
Further investigation into how the priority queue is transforming would
be needed. As it is only documented to return sorted on score, the
quickest solution for now is to drop the secondary sorting.

In the future it would be nice to revisit this to stablise the sort for
the user.