Give up completely when compiled expressions fail
For expressions like `e1 OR e2 OR e3...` we give up compiling the full expression but instead
compile each subexpression (`e1`, `e2`, `e3`, ...). This rarely what we want since it will potentiall
create very many classes each with a small expression. The PR changes that so that expression
compilation is all-or-nothing.