Revert defaulting rel cardinality to 0 (planner integration tests)
While it is tedious to add all relationship counts,
especially with label inference, I have since this was changed often
forgotten to add some relationship count.

Like this we accidentally create more flaky tests. With a count of 0
we get a cost of zero which means that any candidate can win.