PushdownPropertyReads: keep propertyReadOptima in a Map from the beginning.
propertyReadOptima was a Seq before, and turned into a Map after having traversed
the whole logical plan. This is problematic with duplicate entries.
And unfortunately, the `LogicalPlan.foldPlan` framework makes those very easy to occur:

```
.apply()
.|.union() // Acc(propertyReadOptima = Seq(opt1, opt1))
.|.|.z() // Acc(propertyReadOptima = Seq(opt1))
.|.y() // Acc(propertyReadOptima = Seq(opt1))
.x() // Acc(propertyReadOptima = Seq(opt1))
```

This commit is a pure refactoring on the surface of PushdownPropertyReads,
but it builds a `Map[Id, Set[PushDownProperty]]` from the beginning,
thereby avoiding any duplicate values building up.

It also applies the `propertyWithName` function to each `PushDownProperty` directly,
instead of as a 2nd pass over all entries.