Dont call systemdb from dbms runtime repository in enterprise ()
The reason for this is to avoid calling txs to the system database. When running in an enterprise mode it is allowed for the system database to sometimes not be available. For example this is true when there is a store copy on the member (the database is still available from the cluster point of view). We therefore already have a model where updates to system database are cached and notified through events, which is a safe model for system.

This PR removes any transactions against system db and instead relies on the snapshots.